[feature] 关闭调试输出
