+++ lib/src/page/home/view.dart
@@ -5,7 +5,6 @@ import 'package:Cruise/src/page/home/home_model.dart';
 import 'package:fish_redux/fish_redux.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-
 import 'state.dart';
 
 Widget buildView(HomeState state, Dispatch dispatch, ViewService viewService) {
