+++ test/widget_test.dart
@@ -5,7 +5,7 @@
 // gestures. You can also use WidgetTester to find child src.widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
-import 'package:Cruise/src/widgets/CruiseApp.dart';
+import 'package:Cruise/src/widgets/cruise_app.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
